fix(doom/f_wipe): compute wipe_initColorXForm copy size in size_t (CodeQL, alert #7)

+++ kernel/doom_src/f_wipe.c
@@ -67,7 +67,7 @@ wipe_initColorXForm
   int	height,
   int	ticks )
 {
-    memcpy(wipe_scr, wipe_scr_start, width*height);
+    memcpy(wipe_scr, wipe_scr_start, (size_t)width*height);   // 64-bit multiply before widening to size_t
     return 0;
 }
